[feat] entitiy 컬럼 수정

+++ src/main/java/com/my/meetingroom/repository/MeetingroomRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.my.meetingroom.entity.MeetingReservationEntity;
 
-public interface MeetingroomRepository extends JpaRepository<MeetingReservationEntity, Long>{
+public interface MeetingroomRepository extends JpaRepository<MeetingReservationEntity, Integer>{
 
 }
